Move amqp/rabbit server request implementation to separate file
