+++ server/amqp_request.go
@@ -1,0 +1,68 @@
+package server
+
+import (
+	"strings"
+
+	"github.com/streadway/amqp"
+	"golang.org/x/net/context"
+)
+
+type AMQPRequest struct {
+	context.Context
+	body     interface{}
+	delivery *amqp.Delivery
+}
+
+func NewAMQPRequest(delivery *amqp.Delivery) *AMQPRequest {
+	return &AMQPRequest{
+		delivery: delivery,
+	}
+}
+
+// RabbitMQ / AMQP fields
+
+func (r *AMQPRequest) Service() string {
+	routingKey := r.RoutingKey()
+	lastDotIndex := strings.LastIndex(routingKey, ".")
+	if lastDotIndex == -1 {
+		return routingKey
+	}
+	return routingKey[:lastDotIndex]
+}
+
+func (r *AMQPRequest) Endpoint() string {
+	routingKey := r.RoutingKey()
+	lastDotIndex := strings.LastIndex(routingKey, ".")
+	if lastDotIndex == -1 {
+		return ""
+	}
+	return routingKey[lastDotIndex+1:]
+}
+
+func (r *AMQPRequest) Payload() []byte {
+	return r.delivery.Body
+}
+
+func (r *AMQPRequest) Body() interface{} {
+	return r.body
+}
+
+func (r *AMQPRequest) SetBody(body interface{}) {
+	r.body = body
+}
+
+func (r *AMQPRequest) CorrelationID() string {
+	return r.delivery.CorrelationId
+}
+
+func (r *AMQPRequest) ReplyTo() string {
+	return r.delivery.ReplyTo
+}
+
+func (r *AMQPRequest) RoutingKey() string {
+	return r.delivery.RoutingKey
+}
+
+func (r *AMQPRequest) Interface() interface{} {
+	return r.delivery
+}
